+++ src/js/squadSpawnGroup.js
@@ -1,6 +1,6 @@
 import { Marker, Icon, Browser, DomEvent } from "leaflet";
 import { App } from "../app.js";
-import tippy from "tippy.js";
+import tippy, { sticky } from "tippy.js";
 import i18next from "i18next";
 
 
@@ -57,6 +57,7 @@ export const squadSpawnGroup = Marker.extend({
             delay: 200,
             placement: "top",
             sticky: true,
+            plugins: [sticky],
             duration: 0,
             allowHTML: true,
             interactive: true,
@@ -81,7 +82,7 @@ export const squadSpawnGroup = Marker.extend({
         if (this.data.team === "Team One" || this.data.team === "Team Neutral") {
             if (this.layer.factions?.FACTION1_SELECTOR.val() && App.userSettings.enableFactions && process.env.DISABLE_FACTIONS != "true") {
                 team = this.layer.factions.FACTION1_SELECTOR.val();
-                teamName = i18next.t(`factions:${this.layer.factions.FACTION1_SELECTOR.val()}`);
+                teamName = i18next.t(`factions:${team.replace(/^SU_/, "").replace(/-\d+$/, "")}`);
             } 
             else {
                 team = "Team1";
@@ -91,7 +92,7 @@ export const squadSpawnGroup = Marker.extend({
         else {
             if (this.layer.factions?.FACTION2_SELECTOR.val() && App.userSettings.enableFactions && process.env.DISABLE_FACTIONS != "true") {
                 team = this.layer.factions.FACTION2_SELECTOR.val();
-                teamName = i18next.t(`factions:${this.layer.factions.FACTION2_SELECTOR.val()}`);
+                teamName = i18next.t(`factions:${team.replace(/^SU_/, "").replace(/-\d+$/, "")}`);
             } 
             else {
                 team = "Team2";
@@ -176,7 +177,7 @@ export const squadSpawnGroup = Marker.extend({
     _handleContextMenu(event) {
         DomEvent.preventDefault(event);
 
-        if (!this.temporary) return; 
+        if (!this.temporary) return;
 
         // clean up tooltip if needed
         if (this.tippy && this.tippy.destroy) {